requirements.txt
asgiref==3.8.1
bleak==0.22.3
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
diff-match-patch==20230430
distlib==0.3.9
Django==5.1.1
django-debug-toolbar==4.4.6
django-import-export==4.1.1
django-jazzmin==3.0.1
django-unfold==0.40.0
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
network==0.1
packaging==24.1
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.2
platformdirs==4.3.6
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg2==2.9.10
ptyprocess==0.7.0
pyproject_hooks==1.2.0
pyserial==3.5
pywin32-ctypes==0.2.3
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==58.0.0
shellingham==1.5.4
sqlparse==0.5.1
tablib==3.5.0
tomlkit==0.13.2
trove-classifiers==2024.10.13
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.1.3
winrt-runtime==2.2.0
winrt-Windows.Devices.Bluetooth==2.2.0
winrt-Windows.Devices.Bluetooth.Advertisement==2.2.0
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==2.2.0
winrt-Windows.Devices.Enumeration==2.2.0
winrt-Windows.Foundation==2.2.0
winrt-Windows.Foundation.Collections==2.2.0
winrt-Windows.Storage.Streams==2.2.0
